ok remove copied code
